+++ apps/hops-bestilling/src/main/kotlin/no/nav/helse/hops/infrastructure/Configuration.kt
@@ -16,12 +16,4 @@ object Configuration {
     data class FhirMessaging(
         val endpoint: String,
     )
-
-    data class FhirServer(
-        val baseUrl: String,
-        val discoveryUrl: String,
-        val clientId: String,
-        val clientSecret: String,
-        val scope: String,
-    )
 }
