+++ apps/hops-bestilling/src/main/kotlin/no/nav/helse/hops/infrastructure/KoinBootstrapper.kt
@@ -1,19 +1,13 @@
 package no.nav.helse.hops.infrastructure
 
-import ca.uhn.fhir.context.FhirContext
-import ca.uhn.fhir.context.FhirVersionEnum
-import ca.uhn.fhir.rest.api.EncodingEnum
-import ca.uhn.fhir.rest.client.api.IGenericClient
-import ca.uhn.fhir.rest.client.api.ServerValidationModeEnum
-import com.sksamuel.hoplite.ConfigLoader
+import no.nav.helse.hops.FhirClientFactory
 import no.nav.helse.hops.domain.BestillingConsumerJob
 import no.nav.helse.hops.domain.FhirMessageBus
 import no.nav.helse.hops.domain.FhirRepository
 import no.nav.helse.hops.domain.FhirRepositoryImpl
 import no.nav.helse.hops.domain.FhirResourceValidator
+import no.nav.helse.hops.hoplite.loadConfigOrThrow
 import no.nav.helse.hops.koin.singleClosable
-import no.nav.helse.hops.security.fhir.OauthRequestInterceptor
-import no.nav.helse.hops.security.oauth.OAuth2ClientFactory
 import org.koin.dsl.module
 import org.slf4j.Logger
 import org.slf4j.LoggerFactory
@@ -23,17 +17,17 @@ object KoinBootstrapper {
         data class ConfigRoot(
             val kafka: Configuration.Kafka,
             val fhirMessaging: Configuration.FhirMessaging,
-            val fhirServer: Configuration.FhirServer
+            val fhirServer: FhirClientFactory.Config
         )
 
-        single { ConfigLoader().loadConfigOrThrow<ConfigRoot>("/application.conf") }
+        single { loadConfigOrThrow<ConfigRoot>() }
         single { get<ConfigRoot>().kafka }
         single { get<ConfigRoot>().fhirMessaging }
         single { get<ConfigRoot>().fhirServer }
 
         single<FhirResourceValidator> { FhirResourceValidatorHapi }
         single<FhirMessageBus> { FhirMessageBusKafka(get(), get(), get()) }
-        single { createHapiFhirClient(get()) }
+        single { FhirClientFactory.create(get()) }
         single<FhirRepository> { FhirRepositoryImpl(get(), getLogger<FhirRepositoryImpl>()) }
 
         singleClosable { KafkaFactory.createFhirProducer(get()) }
@@ -41,22 +35,6 @@ object KoinBootstrapper {
         singleClosable(createdAtStart = true) {
             BestillingConsumerJob(get(), getLogger<BestillingConsumerJob>(), get(), get(), get())
         }
-    }
-}
-
-private fun createHapiFhirClient(config: Configuration.FhirServer): IGenericClient {
-    val oauthClient = OAuth2ClientFactory.create(
-        config.discoveryUrl, config.clientId, config.clientSecret
-    )
-
-    val interceptor = OauthRequestInterceptor(oauthClient, config.scope)
-
-    // So that we dont start by requesting /metadata.
-    val ctx = FhirContext.forCached(FhirVersionEnum.R4)
-    val factory = ctx.restfulClientFactory.apply { serverValidationMode = ServerValidationModeEnum.NEVER }
-    return factory.newGenericClient(config.baseUrl).apply {
-        registerInterceptor(interceptor)
-        encoding = EncodingEnum.JSON
     }
 }
