+++ apps/hops-hendelser/src/test/kotlin/no/nav/helse/hops/integrationTests/ApplicationTest.kt
@@ -16,7 +16,7 @@ class ApplicationTest {
     @Test
     fun `my test example`() {
         withHopsTestApplication {
-            Thread.sleep(100000)
+            Thread.sleep(1000)
         }
     }
