+++ apps/hops-oppslag/build.gradle.kts
@@ -33,9 +33,7 @@ dependencies {
     implementation("io.insert-koin:koin-ktor:3.0.1")
     implementation("io.ktor:ktor-metrics-micrometer:$ktorVersion")
     implementation("io.micrometer:micrometer-registry-prometheus:1.6.6")
-    runtimeOnly("ca.uhn.hapi.fhir:hapi-fhir-client:$hapiVersion")
     runtimeOnly("ch.qos.logback:logback-classic:1.2.3")
-    runtimeOnly("io.ktor:ktor-client-cio:$ktorVersion")
     runtimeOnly("io.ktor:ktor-server-netty:$ktorVersion")
     runtimeOnly("net.logstash.logback:logstash-logback-encoder:6.6")
     testImplementation(kotlin("test-junit5"))
