+++ libs/hops-common-core/src/main/kotlin/no/nav/helse/hops/security/oauth/OAuth2ClientFactory.kt
@@ -4,7 +4,7 @@ import com.fasterxml.jackson.annotation.JsonInclude
 import com.fasterxml.jackson.databind.DeserializationFeature
 import com.nimbusds.oauth2.sdk.auth.ClientAuthenticationMethod
 import io.ktor.client.HttpClient
-import io.ktor.client.engine.cio.CIO
+import io.ktor.client.engine.java.Java
 import io.ktor.client.features.json.JacksonSerializer
 import io.ktor.client.features.json.JsonFeature
 import no.nav.security.token.support.client.core.ClientAuthenticationProperties
@@ -25,7 +25,7 @@ object OAuth2ClientFactory {
         return OAuth2Client(createHttpClient(), wellKnownUrl, clientAuth)
     }
 
-    private fun createHttpClient() = HttpClient(CIO) {
+    private fun createHttpClient() = HttpClient(Java) {
         install(JsonFeature) {
             serializer = JacksonSerializer {
                 configure(DeserializationFeature.FAIL_ON_UNKNOWN_PROPERTIES, false)
