+++ libs/hops-common-fhir/build.gradle.kts
@@ -17,7 +17,10 @@ tasks {
 }
 
 dependencies {
+    val hapiVersion = "5.3.3"
+
     api(project(":libs:hops-common-core"))
-    api("ca.uhn.hapi.fhir:hapi-fhir-base:5.3.3")
+    api("ca.uhn.hapi.fhir:hapi-fhir-base:$hapiVersion")
     implementation("org.jetbrains.kotlinx:kotlinx-coroutines-jdk8:1.4.3")
+    runtimeOnly("ca.uhn.hapi.fhir:hapi-fhir-client:$hapiVersion")
 }
